Create README.md
The Competition is conducted by Analytics Vidhya named Janatahack - Healthcare Analytics II. The task is to accurately predict the Length of Stay for each patient on case by case basis so that the Hospitals can use this information for optimal resource allocation and better functioning. The length of stay is divided into 11 different classes ranging from 0-10 days to more than 100 days. I tried almost all classifier algorithms for this Multioutput problem to get better accuracy. I achieved an accuracy of around 42.867% using LightGBM which is a very powerful ensemble technique.